@@ -3,7 +3,7 @@ class Plc::EnrollmentEvaluationsController < ApplicationController
 
   def preview_assignments
     if @enrollment_unit_assignment.module_assignment_for_type(Plc::LearningModule::CONTENT_MODULE).nil? && @enrollment_unit_assignment.module_assignment_for_type(Plc::LearningModule::PRACTICE_MODULE).nil?
-      modules_to_enroll_in = Plc::LearningModule.find(@enrollment_unit_assignment.plc_course_unit.determine_preferred_learning_modules(current_user))
+      modules_to_enroll_in = @enrollment_unit_assignment.plc_course_unit.determine_preferred_learning_modules(current_user)
       @enrollment_unit_assignment.enroll_user_in_unit_with_learning_modules(modules_to_enroll_in)
     end
 
