@@ -11,7 +11,7 @@ Dashboard::Application.configure do
   config.eager_load = false
 
   # Always reload static js and css.
-  config.static_cache_control = 'must-revalidate, max-age=0'
+  config.public_file_server.headers = { 'Cache-Control' => 'must-revalidate, max-age=0' }
 
   # Show full error reports
   config.consider_all_requests_local = true
@@ -41,6 +41,8 @@ Dashboard::Application.configure do
   # This option may cause significant delays in view rendering with a large
   # number of complex assets.
 #  config.assets.debug = true
+
+  config.assets.quiet = true
 
   # Whether or not to display pretty apps (formerly called blockly).
   config.pretty_apps = true
