@@ -25,8 +25,8 @@ Dashboard::Application.configure do
   # config.action_dispatch.rack_cache = true
 
   # Serve static resources - these will be cached by Varnish (or a CDN)
-  config.serve_static_files = true
-  config.static_cache_control = "public, max-age=86400, s-maxage=43200"
+  config.public_file_server.enabled = true
+  config.public_file_server.headers = { 'Cache-Control' => "public, max-age=86400, s-maxage=43200" }
 
   # Compress JavaScripts and CSS.
   config.assets.js_compressor = :uglifier
