@@ -18,8 +18,8 @@ Dashboard::Application.configure do
   config.use_schema_cache_dump = true
 
   # Configure static asset server for tests with Cache-Control for performance.
-  config.serve_static_files = true
-  config.static_cache_control = "public, max-age=3600, s-maxage=1800"
+  config.public_file_server.enabled = true
+  config.public_file_server.headers = { 'Cache-Control' => "public, max-age=3600, s-maxage=1800" }
 
   # test environment should use precompiled digested assets like production,
   # unless it's being used for unit tests.
@@ -40,9 +40,14 @@ Dashboard::Application.configure do
     config.assets.version = '1.0'
   end
 
+  config.assets.quiet = true
+
   # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
+
+  # Disable Rails.cache when running unit tests.
+  config.cache_store = :memory_store, { size: 64.megabytes } if ci_test
 
 #  config.action_mailer.raise_delivery_errors = true
 #  config.action_mailer.delivery_method = :smtp
