@@ -2,11 +2,11 @@ require 'test_helper'
 
 class CachingTest < ActionDispatch::IntegrationTest
   def setup
+    Script.stubs(:should_cache?).returns true
     Script.clear_cache
     # turn on the cache (off by default in test env so tests don't confuse each other)
-    Dashboard::Application.config.action_controller.perform_caching = true
-    Dashboard::Application.config.cache_store = :memory_store, { size: 64.megabytes }
-    Rails.application.config.stubs(:levelbuilder_mode).returns false
+    Rails.application.config.action_controller.perform_caching = true
+    Rails.application.config.cache_store = :memory_store, { size: 64.megabytes }
 
     Rails.cache.clear
   end
