@@ -4,9 +4,9 @@ class ExceptionsTest < ActionDispatch::IntegrationTest
   def setup
   end
 
-  test "invalid format for template raises 404 not 500" do
-    get '/', {}, {'HTTP_ACCEPT' => 'image/jpeg'}
+  test "invalid format for template raises 406 not 500" do
+    get '/', params: {}, headers: {'HTTP_ACCEPT' => 'image/jpeg'}
 
-    assert_response 404
+    assert_response 406
   end
 end
