@@ -4,7 +4,6 @@ require 'pathname'
 require 'cdo/aws/s3'
 require 'cdo/hip_chat'
 require 'digest'
-require 'sprockets-derailleur'
 
 module RakeUtils
   def self.system__(command)
@@ -101,7 +100,8 @@ module RakeUtils
   end
 
   def self.nproc
-    SprocketsDerailleur.worker_count
+    # TODO: Replace with system processor count.
+    1
   end
 
   def self.bundle_install(*args)
